+++ doc/CHANGELOG.md
@@ -37,6 +37,7 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#2708` Fix: API key auth does not work for custom_field actions
 * `#2716` Fix: Repository is not auto-created when activated in project settings
 * `#2717` Fix: Multiple journal entries when adding multiple attachments in one change
+* Improve settings performance by caching whether the settings table exists
 
 ## 3.0.0pre27